Move `Mod` into `modu`
