+++ src/core/mod.rs
@@ -1,3 +1,4 @@
 pub mod ident;
+pub mod modu;
 pub mod typ;
 pub mod value;
